Add .gitattributes and normalize line endings
This change adds a .gitattributes file that normalizes line endings in
the repository for all *.vim files.  Line endings will be converted to
the host system format on checkout.  The line endings for all affected
files are also normalized in this commit.  This allows this plugin to
work correctly on OS X and Linux, e.g. with Mono and Unity3D C#
toolchains.

See also: https://help.github.com/articles/dealing-with-line-endings